move usertimings into a computed artifact
